requirements.txt
requests==2.20.0
Pillow==5.4.1
python-dotenv==0.10.1